requirements/base.txt
# Framework
Django==1.8
django-model-utils==2.2
django-environ==0.4.0
psycopg2==2.6
whitenoise==2.0.4
# Apps
#django-scheduler==0.7.4
django-happenings==0.3.3
django-markdown==0.8.4
